+++ src/aggregator/aggregator/generic_elem.go
@@ -617,7 +617,7 @@ func (e *GenericElem) processValueWithAggregationLock(
 		if !e.parsedPipeline.HasRollup {
 			toFlush := make([]transformation.Datapoint, 0, 2)
 			toFlush = append(toFlush, transformation.Datapoint{
-				TimeNanos: int64(timeNanos),
+				TimeNanos: timeNanos,
 				Value:     value,
 			})
 			if extraDp.TimeNanos != 0 {
